% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.htest}
\alias{plot.htest}
\title{Plots htest objects}
\usage{
\method{plot}{htest}(x, legend_pos = "topleft", digits = 4L, quote = TRUE,
  prefix = "", ...)
}
\description{
Plots htest objects.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
sample_test <- cor.test(x, y)
plot(x, y)
plot(sample_test, topleft)
}

