% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{plotest-package}
\alias{plotest-package}
\title{Plot test results}
\description{
Plotest contains a set of \code{\link[graphics]{plot}} methods for R  of classes
\code{htest} or \code{lm} commonly representing results of various statistical
tests.
}
\details{
\tabular{ll}{ Package: \tab plotest\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2014-11-01\cr License: \tab GPL-3 \cr }
}
\keyword{package}

