% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.htest}
\alias{plot.htest}
\title{Plots htest objects}
\usage{
\method{plot}{htest}(x, legend_pos = "topleft", digits = 4L, bty = "n",
  ...)
}
\arguments{
\item{x}{an object of class \code{htest}.}

\item{legend_pos}{a position of the text on the plot. Either matrix or a single keyword.
See Details section of \code{\link[graphics]{legend}}.}

\item{digits}{number of digits.}

\item{bty}{the type of box to be drawn around the plotted test. The allowed values are "o"
(border around plotted test information) and "n" (no border; the default).}

\item{...}{additional arguments sent to \code{\link[graphics]{legend}}.}
}
\description{
Adds htest objects to existing plots.
}
\examples{
x <- 1:20
y <- exp(x)
sample_test <- cor.test(x, y)
plot(x, y)
plot(sample_test, "topleft")
}

