% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.htest}
\alias{plot.htest}
\title{Plots htest objects}
\usage{
\method{plot}{htest}(x, legend_pos = "topleft", digits = 4L, bty = "n",
  n_char = 40, exdent = 2, ...)
}
\arguments{
\item{x}{an object of class \code{htest}.}

\item{legend_pos}{a position of the text on the plot. Either matrix or a single keyword.
See Details section of \code{\link[graphics]{legend}}.}

\item{digits}{number of digits.}

\item{bty}{the type of box to be drawn around the plotted test. The allowed values are "o"
(border around plotted test information) and "n" (no border; the default).}

\item{n_char}{\code{integer} representing maximum number of character in single line
of the output.}

\item{...}{additional arguments sent to \code{\link[graphics]{legend}}.}

\item{extend}{a non-negative integer specifying the indentation of subsequent lines in
paragraphs.}
}
\description{
Adds htest objects to existing plots.
}
\details{
Longer lines are splitted on whitespaces using
\code{\link[base]{strwrap}} function.
}
\examples{
#Example 1: test for correlation between paired samples
x <- 1:20
y <- exp(x)
sample_cortest <- cor.test(x, y)
plot(x, y, type = "b")
plot(sample_cortest, "topleft")

#Example 2: Pearson's Chi-squared Test for Count Data
x <- matrix(c(12, 5, 7, 7), ncol = 2)
sample_chisqtest <- chisq.test(x, y)
barplot(x, beside = TRUE)
plot(sample_chisqtest, "topright")
}

